Change method to check for table existence
The method ```fruit_table_exists?(db)``` returns an empty array if the table does not exist. 

This is then used on line 41 to check whether the result is 'truthy'. 

It passes even if the table doesn't exist, I think that an empty array must be 'truthy', so I've changed the method to return the first element of the array which will be nil if it is empty.